2089 overwritten error log (#2140)
* Add code to create log if it does not already exist.

+++ arches/app/utils/data_management/resources/formats/format.py
@@ -137,18 +137,21 @@ class Reader(object):
         else:
             print _("***** Errors occured during import. Some data may not have been imported. For more information, check resource import error log: arches/logs/resource_import.log")
             log_nums = [0]
-            if os.path.getsize('arches/logs/resource_import.log')/1000000 > 5:
-                for file in os.listdir('arches/logs',):
-                    try:
-                        log_nums.append(int(file.split('.')[-1]))
-                    except:
-                        pass
+            if os.path.isfile('arches/logs/resource_import.log'):
+                if os.path.getsize('arches/logs/resource_import.log')/1000000 > 5:
+                    for file in os.listdir('arches/logs',):
+                        try:
+                            log_nums.append(int(file.split('.')[-1]))
+                        except:
+                            pass
 
-                archive_log_num = str(max(log_nums) + 1)
-                shutil.copy2('arches/logs/resource_import.log', 'arches/logs/resource_import.log' + '.' + archive_log_num)
-                f = open('arches/logs/resource_import.log', 'w')
+                    archive_log_num = str(max(log_nums) + 1)
+                    shutil.copy2('arches/logs/resource_import.log', 'arches/logs/resource_import.log' + '.' + archive_log_num)
+                    f = open('arches/logs/resource_import.log', 'w')
+                else:
+                    f = open('arches/logs/resource_import.log', 'a')
             else:
-                f = open('arches/logs/resource_import.log', 'a')
+                f = open('arches/logs/resource_import.log', 'w')
 
             for error in self.errors:
                 timestamp = (datetime.datetime.now() - datetime.timedelta(hours=2)).strftime('%a %b %d %H:%M:%S %Y')
